+++ site/components/molecules/other-referred.tsx
@@ -58,7 +58,7 @@ const OtherReferredMolecule = () => {
 
 	return (
 		<>
-			<Typography variant="h5" component="h3" gutterBottom sx={{ mt: 2 }}>
+			<Typography id="other" variant="h5" component="h3" gutterBottom sx={{ mt: 2 }}>
 				Other referred publications
 			</Typography>