+++ CHANGES.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 - Added `Technical reports` molecule
+- Created `menu` molecule
 
 ## 0.4.0
 - Added `Other referred` molecule